+++ src/methods/search.js
@@ -1,5 +1,5 @@
-const { parseOptions, requests } = require('../utils/utils.js');
-const parse = require('../parser/main.js');
+const { getProp, parseOptions, requests } = require('../utils/utils.js');
+const { parse } = require('../parser/main.js');
 
 async function search(searchString, options){
 	if(!searchString){
@@ -12,52 +12,39 @@ async function search(searchString, options){
 	const search_query = encodeURIComponent(
 		searchString.trim().split(/\s+/).join('+')
 	);
-
-	let body = await requests.fetch(
+	const body = await requests.fetch(
 		`https://www.youtube.com/results?search_query=${search_query}`, 
 		options
 	).text();
-	let data = requests.getData(body, 1), ytcfg = requests.getData(body, 3);
 
-	const contents = data.contents
-		.twoColumnSearchResultsRenderer
-		.primaryContents.sectionListRenderer.contents;
-		
-	if(
-		contents.length === 1 && 
-		contents[0].itemSectionRenderer && 
-		contents[0].itemSectionRenderer.contents[0].backgroundPromoRenderer
-	){
-		if(options.raw) return { initialData: data, ytcfg };
-		return null;
+	const data = requests.getData(body, 1);
+	let results = getProp(data, 'contents.twoColumnSearchResultsRenderer.primaryContents.sectionListRenderer.contents');
+
+	let continuationItem = results.pop(); 
+	results = getProp(results.pop(), 'itemSectionRenderer.contents');
+
+	if(!results) return null;
+
+	if(results.length < options.quantity){
+		const ytcfg = requests.getData(body, 3);
+
+		while(results.length < options.quantity){
+			let continuation = await requests.getContinuation(continuationItem, ytcfg, options);
+			let items = getProp(continuation, 'onResponseReceivedCommands.0.appendContinuationItemsAction.continuationItems');
+
+			results.push(...items[0].itemSectionRenderer.contents);
+			
+			continuationItem = items[1];
+			if(!continuation) break;
+		}
 	}
-
-
-	let continuationItem = contents.pop();
-	let results = contents.find(a => 
-		a.itemSectionRenderer && a.itemSectionRenderer.contents.length > 2
-	).itemSectionRenderer.contents;
-	
-	while(results.length < options.quantity){
-		let continuation = await requests.getContinuation(continuationItem, ytcfg, options);
-		
-		let items = continuation.onResponseReceivedCommands[0]
-			.appendContinuationItemsAction.continuationItems;
-
-		results.push(...items[0].itemSectionRenderer.contents);
-		
-		continuationItem = items[1];
-		if(!continuationItem) break;
-	}
-	
-
-	if(options.raw) return { initialData: data, ytcfg, items: results };
 
 	return {
-		searchQuery: search_query.trim(),
+		searchQuery: searchString.trim(),
 		estimatedResults: Number(data.estimatedResults),
-		results: results.map(parse).filter(x => x)
+		refinements: data.refinements,
+		results: results.map(parse).filter(x => x),
 	};
 }
-
+	
 module.exports = search;