+++ src/parser/main.js
@@ -1,207 +1,93 @@
-const {
-	Thumbnails, Duration, Views,
-	parseText, extractInt
-} = require('./structures.js');
+const Utils = require('./utils.js');
 
 //#region video
-function videoPrimaryInfoRenderer({ videoPrimaryInfoRenderer }){
-	const [likes, dislikes] = videoPrimaryInfoRenderer.sentimentBar
-		.sentimentBarRenderer.tooltip.split(' / ').map(extractInt);
+function compactVideoRenderer({ compactVideoRenderer, endScreenVideoRenderer }){
+	let videoData = compactVideoRenderer || endScreenVideoRenderer;
 
-	return {
-		name: parseText(videoPrimaryInfoRenderer.title),
-		views: new Views(videoPrimaryInfoRenderer.viewCount),
-
-		likes, dislikes,
-
-		uploadDateLabel: parseText(videoPrimaryInfoRenderer.dateText),
-	};
-}
-
-function videoSecondaryInfoRenderer({ videoSecondaryInfoRenderer }){
-	return {
-		owner: parse(videoSecondaryInfoRenderer.owner),
-		description: parseText(videoSecondaryInfoRenderer.description)
-	};
-}
-
-function compactVideoRenderer({ compactVideoRenderer }){
 	let data = {
-		name: parseText(compactVideoRenderer.title),
-		ID: compactVideoRenderer.videoId,
-		URL: `https://www.youtube.com/watch?v=${compactVideoRenderer.videoId}`,
+		name: Utils.parseText(videoData.title).toString(),
+		ID: videoData.videoId,
+		URL: `https://www.youtube.com/watch?v=${videoData.videoId}`,
 		type: 'video',
 
-		views: new Views(compactVideoRenderer),
-		thumbnails: new Thumbnails(compactVideoRenderer.thumbnail),
+		views: new Utils.Views(videoData),
+		thumbnails: new Utils.Thumbnails(videoData.thumbnail),
 
-		publishedDate: parseText(compactVideoRenderer.publishedTimeText),
+		publishedDate: Utils.parseText(videoData.publishedTimeText).toString(),
 
-		channel: bylineText(compactVideoRenderer)
+		owner: bylineText(videoData),
 	};
 
-	if(compactVideoRenderer.lengthText){
-		data.duration = new Duration(compactVideoRenderer);
-	}
+	if(videoData.lengthText) data.duration = new Utils.Duration(videoData);
 
 	return data;
 }
 
-function compactRadioRenderer({ compactRadioRenderer }){
+function compactRadioRenderer({ compactRadioRenderer, endScreenPlaylistRenderer }){
+	let data = compactRadioRenderer || endScreenPlaylistRenderer;
+
 	return {
-		name: parseText(compactRadioRenderer.title),
-		ID: compactRadioRenderer.playlistId,
-		URL: compactRadioRenderer.shareUrl || `https://www.youtube.com/playlist?list=${compactRadioRenderer.playlistId}`,
+		name: Utils.parseText(data.title).toString(),
+		ID: data.playlistId,
+		URL: data.shareUrl || `https://www.youtube.com/playlist?list=${data.playlistId}`,
 		type: 'playlist',
 
-		thumbnails: new Thumbnails(compactRadioRenderer.thumbnail),
+		thumbnails: new Utils.Thumbnails(data.thumbnail),
 
-		owner: bylineText(compactRadioRenderer),
-	};
-}
-
-function watchNextEndScreenRenderer({ watchNextEndScreenRenderer }){
-	if(!watchNextEndScreenRenderer.results) return null;
-
-	return {
-		title: parseText(watchNextEndScreenRenderer.title),
-		items: watchNextEndScreenRenderer.results.map(parse)
-	};
-}
-
-let endScreenVideoRenderer = ({ endScreenVideoRenderer }) => {
-	return compactVideoRenderer({
-		compactVideoRenderer: endScreenVideoRenderer
-	});
-};
-
-let endScreenPlaylistRenderer = ({ endScreenPlaylistRenderer }) => {
-	return compactRadioRenderer({
-		compactRadioRenderer: endScreenPlaylistRenderer
-	});
-};
-
-function playlist({ playlist }){//playlist where the video is
-	return {
-		ID: playlist.playlistId,
-		title: parseText(playlist.titleText),
-
-		owner: bylineText(playlist),
-
-		videoQuantity: playlist.totalVideos,
-
-		videos: playlist.videos.map(parse)
-	};
-}
-
-function playlistPanelVideoRenderer({ playlistPanelVideoRenderer }){ //playlist video
-	return {
-		name: parseText(playlistPanelVideoRenderer.title),
-		ID: playlistPanelVideoRenderer.videoId,
-		playlistID: playlistPanelVideoRenderer.navigationEndpoint.watchEndpoint.playlistId,
-		URL: `https://www.youtube.com${playlistPanelVideoRenderer.navigationEndpoint.commandMetadata.webCommandMetadata.url}`,
-
-		duration: new Duration(playlistPanelVideoRenderer),
-		views: new Views(playlistPanelVideoRenderer),
-		thumbnails: new Thumbnails(playlistPanelVideoRenderer.thumbnail),
-
-		publishedDate: parseText(playlistPanelVideoRenderer.publishedTimeText),
-
-		channel: bylineText(playlistPanelVideoRenderer)
-	};
-}
-//#endregion
-
-//#region playlist
-function playlistSidebarRenderer({ playlistSidebarRenderer }){
-	let info = parse(playlistSidebarRenderer.items[0]);
-
-	let ownerInfo = parse(playlistSidebarRenderer.items[1].playlistSidebarSecondaryInfoRenderer.videoOwner);
-
-	info.owner = ownerInfo;
-
-	return info;
-}
-
-function playlistSidebarPrimaryInfoRenderer({ playlistSidebarPrimaryInfoRenderer }){
-	let ID = playlistSidebarPrimaryInfoRenderer.navigationEndpoint.watchEndpoint.playlistId;
-
-	return {
-		ID,
-		URL: `https://www.youtube.com/playlist?list=${ID}`,
-		name: parseText(playlistSidebarPrimaryInfoRenderer.title),
-
-		videoQuantity: extractInt(playlistSidebarPrimaryInfoRenderer.stats[0]),
-		views: extractInt(playlistSidebarPrimaryInfoRenderer.stats[1]),
-		lastUpdate: parseText(playlistSidebarPrimaryInfoRenderer.stats[2]),
-
-		description: parseText(playlistSidebarPrimaryInfoRenderer.description),
-
-		thumbnails: new Thumbnails(
-			playlistSidebarPrimaryInfoRenderer.thumbnailRenderer
-				.playlistVideoThumbnailRenderer.thumbnail
-		)
-	};
-}
-
-function playlistVideoRenderer({ playlistVideoRenderer }){
-	return {
-		title: parseText(playlistVideoRenderer.title),
-
-		ID: playlistVideoRenderer.videoId,
-		URL: `https://www.youtube.com/watch?v=${playlistVideoRenderer.videoId}`,
-		
-		index: extractInt(playlistVideoRenderer.index),
-
-		thumbnails: new Thumbnails(playlistVideoRenderer.thumbnail),
-		duration: new Duration(playlistVideoRenderer),
-
-		channel: bylineText(playlistVideoRenderer)
+		owner: bylineText(data),
 	};
 }
 //#endregion
 
 //#region search
-function videoRenderer({ videoRenderer }){
+function videoRenderer({ videoRenderer, promotedVideoRenderer }){
+	let videoData = videoRenderer || promotedVideoRenderer;
+
 	let data = {
-		ID: videoRenderer.videoId,
-		URL: `https://www.youtube.com/watch?v=${videoRenderer.videoId}`,
+		ID: videoData.videoId,
+		URL: `https://www.youtube.com/watch?v=${videoData.videoId}`,
 		type: 'video',
 		
-		title: parseText(videoRenderer.title),
-		description: parseText(
-			videoRenderer.descriptionSnippet
+		title: Utils.parseText(videoData.title).toString(),
+		description: Utils.parseText(
+			videoData.descriptionSnippet
 		),
 
-		thumbnails: new Thumbnails(videoRenderer.thumbnail),
-		views: new Views(videoRenderer),
+		thumbnails: new Utils.Thumbnails(videoData.thumbnail),
+		views: new Utils.Views(videoData),
 
-		channel: bylineText(videoRenderer),
+		owner: bylineText(videoData),
 
-		publishedTime: parseText(videoRenderer.publishedTimeText),
+		publishedTime: Utils.parseText(videoData.publishedTimeText).toString(),
 	};
 
-	if(videoRenderer.lengthText){
-		data.duration = new Duration(videoRenderer);
+	if(videoData.lengthText){
+		data.duration = new Utils.Duration(videoData);
 	}
 
-	data.channel.thumbnails = new Thumbnails(
-		videoRenderer
-			.channelThumbnailSupportedRenderers
-			.channelThumbnailWithLinkRenderer
-			.thumbnail
-	);
+	if(videoData.channelThumbnailSupportedRenderers){
+		data.owner.thumbnails = new Utils.Thumbnails(
+			videoData
+				.channelThumbnailSupportedRenderers
+				.channelThumbnailWithLinkRenderer
+				.thumbnail
+		);
+	}
+
+	if(promotedVideoRenderer){
+		data.AD = promotedVideoRenderer.navigationEndpoint.urlEndpoint.url;
+	}
 
 	return data;
 }
 
 function childVideoRenderer({ childVideoRenderer }){
 	return {
-		title: parseText(childVideoRenderer.title),
+		title: Utils.parseText(childVideoRenderer.title).toString(),
 		ID: childVideoRenderer.videoId,
 		URL: `https://www.youtube.com/watch?v=${childVideoRenderer.videoId}`,
 
-		duration: new Duration(childVideoRenderer),
+		duration: new Utils.Duration(childVideoRenderer),
 	};
 }
 
@@ -209,7 +95,7 @@ function playlistRenderer({ playlistRenderer }){
 	return {
 		ID: playlistRenderer.playlistId,
 		URL: `https://www.youtube.com/playlist?list=${playlistRenderer.playlistId}`,
-		title: parseText(playlistRenderer.title),
+		title: Utils.parseText(playlistRenderer.title).toString(),
 
 		type: 'playlist',
 		videoCount: Number(playlistRenderer.videoCount),
@@ -218,7 +104,7 @@ function playlistRenderer({ playlistRenderer }){
 
 		owner: bylineText(playlistRenderer),
 
-		thumbnails: new Thumbnails(
+		thumbnails: new Utils.Thumbnails(
 			playlistRenderer.thumbnailRenderer.playlistVideoThumbnailRenderer.thumbnail
 		),
 	};
@@ -230,88 +116,49 @@ function channelRenderer({ channelRenderer }){
 		URL: `https://www.youtube.com/channel/${channelRenderer.channelId}`,
 		type: 'channel',
 
-		name: parseText(channelRenderer.title),
+		name: Utils.parseText(channelRenderer.title).toString(),
 
-		description: parseText(channelRenderer.descriptionSnippet),
+		description: Utils.parseText(channelRenderer.descriptionSnippet),
 		
-		thumbnails: new Thumbnails(channelRenderer.thumbnail),
-		videoCount: extractInt(channelRenderer.videoCountText),
-		subscribers: extractInt(channelRenderer.subscriberCountText)
+		thumbnails: new Utils.Thumbnails(channelRenderer.thumbnail),
+		videoCount: Utils.extractInt(channelRenderer.videoCountText),
+		subscribers: Utils.extractInt(channelRenderer.subscriberCountText),
 	};
 }
 
 function shelfRenderer({ shelfRenderer }){
 	let { 
-		items, collapsedStateButtonText
+		items, collapsedStateButtonText,
 	} = shelfRenderer.content.verticalListRenderer;
 
 	return {
-		title: parseText(shelfRenderer.title),
-		label: parseText(collapsedStateButtonText),
+		title: Utils.parseText(shelfRenderer.title).toString(),
+		label: Utils.parseText(collapsedStateButtonText).toString(),
 		type: 'shelf',
-		items: items.map(parse)
-	};
-}
-
-function searchRefinementCardRenderer({ searchRefinementCardRenderer }){
-	return {
-		title: parseText(searchRefinementCardRenderer.query),
-		thumbnails: new Thumbnails(searchRefinementCardRenderer.thumbnail)
+		items: items.map(parse),
 	};
 }
 
 function horizontalCardListRenderer({ horizontalCardListRenderer }){
 	return {
-		title: parseText(
+		title: Utils.parseText(
 			horizontalCardListRenderer.header
 				.richListHeaderRenderer.title
-		),
+		).toString(),
 		type: 'horizontalCardList',
-		items: horizontalCardListRenderer.cards.map(searchRefinementCardRenderer)
-	};
-}
-
-function promotedVideoRenderer({ promotedVideoRenderer }){
-	let data = {
-		ID: promotedVideoRenderer.videoId,
-		URL: `https://www.youtube.com/watch?v=${promotedVideoRenderer.videoId}`,
-		AD: promotedVideoRenderer.navigationEndpoint.urlEndpoint.url,
-		type: 'promotedVideo',
-		
-		title: parseText(promotedVideoRenderer.title),
-		description: parseText(
-			promotedVideoRenderer.descriptionSnippet
+		items: horizontalCardListRenderer.cards.map(
+			({ searchRefinementCardRenderer }) => ({
+				title: Utils.parseText(searchRefinementCardRenderer.query).toString(),
+				thumbnails: new Utils.Thumbnails(searchRefinementCardRenderer.thumbnail),
+			})
 		),
-
-		thumbnails: new Thumbnails(promotedVideoRenderer.thumbnail),
-		views: new Views(promotedVideoRenderer),
-
-		channel: bylineText(promotedVideoRenderer),
-
-		publishedTime: parseText(promotedVideoRenderer.publishedTimeText),
 	};
-
-	if(promotedVideoRenderer.lengthText){
-		data.duration = new Duration(promotedVideoRenderer);
-	}
-
-	if(promotedVideoRenderer.channelThumbnailSupportedRenderers){
-		data.channel.thumbnails = new Thumbnails(
-			promotedVideoRenderer
-				.channelThumbnailSupportedRenderers
-				.channelThumbnailWithLinkRenderer
-				.thumbnail
-		);
-	}
-
-
-	return data;
 }
 
 function searchPyvRenderer({ searchPyvRenderer }){
 	return {
 		type: 'searchAds',
-		items: searchPyvRenderer.ads.map(parse),
+		items: searchPyvRenderer.ads.map(videoRenderer),
 	};
 }
 //#endregion
@@ -321,34 +168,22 @@ function videoOwnerRenderer({ videoOwnerRenderer }){
 	let { browseId, canonicalBaseUrl } = videoOwnerRenderer.navigationEndpoint.browseEndpoint;
 
 	let data = {
-		name: parseText(videoOwnerRenderer.title),
-		thumbnails: new Thumbnails(videoOwnerRenderer.thumbnail),
-
-		channel: {
-			ID: browseId,
-			URL: `https://www.youtube.com/channel/${browseId}`
-		},
-
-		user: {
-			URL: `https://www.youtube.com${canonicalBaseUrl}`
-		}
+		name: Utils.parseText(videoOwnerRenderer.title).toString(),
+		ID: browseId,
+		URL: `https://www.youtube.com/channel/${browseId}`,
+		canonicalURL: `https://www.youtube.com${canonicalBaseUrl}`,
+		thumbnails: new Utils.Thumbnails(videoOwnerRenderer.thumbnail),
 	};
-	if(canonicalBaseUrl.startsWith('/user/')){
-		data.user.name = canonicalBaseUrl.slice(6);
-	}	
-	if(canonicalBaseUrl.startsWith('/channel/')){
-		data.user.id = canonicalBaseUrl.slice(9);
-	}
 
 	if(videoOwnerRenderer.subscriberCountText){
-		let normal = parseText(videoOwnerRenderer.subscriberCountText);
+		let normal = Utils.parseText(videoOwnerRenderer.subscriberCountText).toString();
 
 		data.subscribers = {
 			normal,
-			number: extractInt(normal),
+			number: Utils.extractInt(normal),
 			toString(){
 				return normal;
-			}
+			},
 		};
 	}
 
@@ -358,42 +193,21 @@ function videoOwnerRenderer({ videoOwnerRenderer }){
 function bylineText({ longBylineText, shortBylineText }){//channel/owner
 	let obj = longBylineText || shortBylineText;
 
-	let text = parseText(obj);
+	let text = Utils.parseText(obj);
 	if(!obj.runs) return text;
-	let endpoint = obj.runs.find(
-		obj => obj.navigationEndpoint
-	);
+	let endpoint = obj.runs.find(obj => obj.navigationEndpoint);
 
-	if(!endpoint) return text;
+	if(!endpoint) return text.toString();
 	endpoint = endpoint.navigationEndpoint.browseEndpoint;
 
 	let data = {
-		name: text,
+		name: text.toString(),
 		ID: endpoint.browseId,
-		URL: `https://www.youtube.com/channel/${endpoint.browseId}`
+		URL: `https://www.youtube.com/channel/${endpoint.browseId}`,
 	};
 
 	if(endpoint.canonicalBaseUrl){
 		data.canonicalURL = `https://www.youtube.com${endpoint.canonicalBaseUrl}`;
-	}
-
-	return data;
-}
-
-function alertRenderer({ alertRenderer }){
-	return {
-		type: alertRenderer.type,
-		reason: parseText(alertRenderer.text)
-	};
-}
-
-function playerErrorMessageRenderer({ playerErrorMessageRenderer }){
-	let data = {
-		reason: parseText(playerErrorMessageRenderer.reason),
-		thumbnails: new Thumbnails(playerErrorMessageRenderer.thumbnail)
-	};
-	if(playerErrorMessageRenderer.subreason){
-		data.subreason = parseText(playerErrorMessageRenderer.subreason);
 	}
 
 	return data;
@@ -414,46 +228,37 @@ function playerMicroformatRenderer({ playerMicroformatRenderer }){
 //#endregion
 
 const parsers = {
-	videoPrimaryInfoRenderer,
-	videoSecondaryInfoRenderer,
-	compactVideoRenderer,
-	compactRadioRenderer,
-	playlistPanelVideoRenderer,
-	watchNextEndScreenRenderer,
-	endScreenVideoRenderer,
-	endScreenPlaylistRenderer,
-	playlist,
-	playerMicroformatRenderer,
-
-	playlistSidebarRenderer,
-	playlistSidebarPrimaryInfoRenderer,  
-	playlistVideoRenderer,
+	compactVideoRenderer, 
+	compactRadioRenderer, 
+	endScreenVideoRenderer: compactVideoRenderer,
+	endScreenPlaylistRenderer: compactRadioRenderer,
 
 	channelRenderer,
 	shelfRenderer,
 	videoRenderer,
 	playlistRenderer,
+	
 	horizontalCardListRenderer,
 	searchPyvRenderer,
-	promotedVideoRenderer,
 
 	videoOwnerRenderer,
-	playerErrorMessageRenderer,
-	alertRenderer
+	bylineText,
+	playerMicroformatRenderer,
 };
 
 function parse(obj){
 	let key = Object.keys(obj)[0];
 
-	let parser = parsers[key];
-	if(!parser){
-		//console.warn(`Cannot parse: ${key}`);
+	if(!parsers[key]){
 		//require('fs').writeFileSync(`./${key}.json`, JSON.stringify(obj, null, '\t'));
-		if(global['@Fabricio-191{ debugging }']) debugger;
 		return null;
 	}
 
-	return parser(obj);
+	return parsers[key](obj);
 }
 
-module.exports = parse;
+module.exports = {
+	parsers,
+	parse,
+	Utils,
+};