+++ test/test.js
@@ -1,13 +1,12 @@
-/* eslint-disable no-unused-vars */
 /* eslint-disable no-console */
 /* eslint-disable no-unreachable */
 const fs = require('fs');
 
 const { getVideo, getPlaylist, search } = require('../')
 	.setDefaultOptions({ 
-		location: 'AR',
+		location: 'US',
 		language: 'es-419',
-		quantity: 120
+		quantity: 120,
 	});
 
 (async () => {
@@ -16,19 +15,23 @@ const { getVideo, getPlaylist, search } = require('../')
 	console.log('Getting video info');
 	data.video = await getVideo(
 		'https://www.youtube.com/watch?v=H2wCwdHk-ao&list=PLDS0dpumEOi0pu_0pCGqvcaRkxg-o1gqg'
-	);
+	).catch(err => { throw err; });
 
 	console.log('Getting playlist');
 	data.playlist = await getPlaylist(
-		'https://www.youtube.com/watch?v=H2wCwdHk-ao&list=PLDS0dpumEOi0pu_0pCGqvcaRkxg-o1gqg', 
-	);
+		'https://www.youtube.com/watch?v=H2wCwdHk-ao&list=PLDS0dpumEOi0pu_0pCGqvcaRkxg-o1gqg' 
+	).catch(err => { throw err; });
 
 	console.log('Searching...');
-	data.search = await search('Node.js', { quantity: 100 });
+	data.search = await search('Node.js', { quantity: 50 })
+		.catch(err => { throw err; });
 
 	console.log('Works well :D');
-	if(fs.existsSync('./test')) return;
-	fs.writeFileSync('./results.json', JSON.stringify(data, null, '\t'));
+	
+	fs.writeFileSync(
+		fs.existsSync('./test') ? './test/results.json' : './results.json', 
+		JSON.stringify(data, null, '\t')
+	);
 })();
 
 /*